extract text theme
